Rename Project1_Clune.pbix to ExploritoryDataAnalysis.pbix
Summarizes descriptive statistics about the data, including at least one descriptive visualization, such as a box-and-whisker plot or histogram displaying the distribution of a variable
Contains an interactive element that lets a user cut into the data, e.g., an Excel pivot table with slicers
Generates interesting questions about the data and explores those questions with two or more visualizations that highlight an insight about the data, possibly applying transformations to the data. For example, a scatter or line plot showing the relationship between two variables.
Effectively uses of analytical storytelling to convey insight(s), including good practices for visualizations.